Follow move of calibration_file_io into pyrs.utilities
