+++ prototypes/calibration/Quick_Calibration_Corr.py
@@ -13,7 +13,7 @@ import math
 import lmfit
 
 from pyrs.core import reduce_hb2b_pyrs
-from pyrs.core import calibration_file_io
+from pyrs.utilities import calibration_file_io
 from pyrs.core import reductionengine
 from pyrs.core import mask_util
 from mantid.simpleapi import CreateWorkspace, FitPeaks
