+++ prototypes/calibration/Quick_Calibration_PeakFit.py
@@ -15,7 +15,7 @@ import lmfit
 from pyrs.core import pyrscore
 from pyrs.core import reduce_hb2b_pyrs
 from pyrs.core import instrument_geometry
-from pyrs.core import calibration_file_io
+from pyrs.utilities import calibration_file_io
 from pyrs.core import reductionengine
 from pyrs.core import mask_util
 from mantid.simpleapi import CreateWorkspace, FitPeaks
