+++ prototypes/calibration/Quick_Calibration_Quasi_HB2B.py
@@ -11,7 +11,7 @@ import itertools
 import math
 
 from pyrs.core import reduce_hb2b_pyrs
-from pyrs.core import calibration_file_io
+from pyrs.utilities import calibration_file_io
 from pyrs.core import reduction_manager
 from pyrs.core import mask_util
 from mantid.simpleapi import CreateWorkspace, FitPeaks
