+++ pyrs/core/reduce_hb2b_mtd.py
@@ -7,8 +7,8 @@ from mantid.simpleapi import SortXAxis, CreateWorkspace
 from mantid.api import AnalysisDataService as ADS
 from pyrs.utilities import checkdatatypes
 from pyrs.utilities import file_util
-import calibration_file_io
-from pyrs.core.calibration_file_io import ResidualStressInstrumentCalibration
+from pyrs.utilities import calibration_file_io
+from pyrs.utilties.calibration_file_io import ResidualStressInstrumentCalibration
 
 
 def histogram_data(raw_vec_x, raw_vec_y, target_vec_2theta):
