+++ pyrs/interface/manualreductionwindow.py
@@ -5,7 +5,7 @@ except ImportError:
     from PyQt4.QtGui import QMainWindow, QFileDialog, QVBoxLayout
     from PyQt4.uic import loadUi as load_ui
 from pyrs.core.pyrscore import PyRsCore
-from pyrs.core import calibration_file_io
+from pyrs.utilities import calibration_file_io
 from ui.diffdataviews import DetectorView, GeneralDiffDataView
 import os
 import gui_helper
@@ -645,4 +645,3 @@ class ManualReductionWindow(QMainWindow):
         self._core = pyrs_core
 
         return
-
