+++ scripts/reduce_HB2B_prototype.py
@@ -3,7 +3,7 @@ import sys
 import os
 from pyrs.core import reduction_manager
 from pyrs.utilities import checkdatatypes
-from pyrs.core import calibration_file_io
+from pyrs.utilities import calibration_file_io
 from pyrs.core import mask_util
 from matplotlib import pyplot as plt
